Leniwe parsowanie, manager dock widgetow, multiserie
git-svn-id: https://localhost:8443/svn/hm-mde-1@551 fca3399c-2208-499e-aaa1-2cf7fb07c25e

+++ trunk/src/edrutils/src/kinematiclib/JointAnglesCollection.cpp
@@ -91,7 +91,7 @@ void JointAnglesCollection::createQuaternionRepresentation(kinematic::SkeletalDa
 	
 
 	for (unsigned int i = 0; i < jointsData.size(); ++i) {
-		JointAngleChannelPtr channel(new JointAngleChannel(fps)); 
+		JointAngleChannelPtr channel(new JointAngleChannel(static_cast<float>(fps))); 
 		std::string name = haSkeleton->mapJointName(jointsData[i]->name);
 		channel->setName(name);
 		channelsMap[name] = channel;
